requirements.txt
absl-py==2.1.0
accelerate==0.34.2
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==4.1.0
Bottleneck==1.3.7
Brotli==1.0.9
cachetools==5.5.0
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
comm==0.2.1
contourpy==1.2.0
cycler==0.11.0
dataclasses-json==0.6.7
datasets==3.0.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.30.3
dill==0.3.8
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==0.8.3
fastapi==0.115.4
fastjsonschema==2.16.2
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.3
gast==0.4.0
google-auth==2.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.2
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.25.0
idna==3.10
importlib-metadata==7.0.1
importlib_resources==6.4.0
ipykernel==6.28.0
ipython==8.15.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.7.0
joblib==1.4.2
JPype1==1.5.0
json5==0.9.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.0
jupyter_server==2.14.1
jupyter_server_terminals==0.4.4
jupyterlab==4.2.5
jupyterlab-pygments==0.1.2
jupyterlab_server==2.27.3
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
konlpy==0.6.0
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.17
langchain-openai==0.2.8
langchain-text-splitters==0.3.2
langsmith==0.1.140
libclang==18.1.1
lxml==5.3.0
Mako==1.3.5
Markdown==3.7
MarkupSafe==2.1.3
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.6
mecab==0.996.5
mecab-ko-dic-msvc==0.999
mecab-ko-msvc==0.999
mistune==2.0.4
mkl_fft==1.3.10
mkl_random==1.2.7
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.2.2
notebook_shim==0.2.3
numexpr==2.8.7
numpy==1.26.4
oauthlib==3.2.2
openai==1.54.3
openpyxl==3.1.5
opt-einsum==3.3.0
optuna==4.0.0
orjson==3.10.11
overrides==7.4.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
platformdirs==3.10.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.43
protobuf==3.19.6
psutil==5.9.0
pure-eval==0.2.2
py-hanspell==1.1
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.6.1
Pygments==2.15.1
PyMySQL==1.1.1
pyngrok==7.2.1
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
pywin32==305.1
pywinpty==2.0.10
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.30.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.2
setuptools==75.1.0
sip==6.7.12
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.35
stack-data==0.2.0
starlette==0.41.2
sympy==1.12
tenacity==9.0.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.17.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.2.1
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchvision==0.19.1+cu118
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
typing_extensions==4.11.0
typing-inspect==0.9.0
tzdata==2023.3
unicodedata2==15.1.0
urllib3==2.2.2
uvicorn==0.32.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.44.0
win-inet-pton==1.1.0
wordcloud==1.9.4
wrapt==1.16.0
xxhash==3.5.0
yarl==1.13.1
zipp==3.17.0
